NetworkControl: implementation changed to support COMRPC based interfaces
